# Speed up rebuilding classic hex from string