Using string stream is extremely slow.
Instead this just iterates over the string and decodes each hex character as we go.

Benchmark results from running google benchmark on the side:
```
------------------------------------------------------------
Benchmark                  Time             CPU   Iterations
------------------------------------------------------------
BM_SoleStream            623 ns          623 ns      1117378
BM_SoleIter             91.2 ns         91.2 ns      7631954
```

Also added a benchmark to this repo, here are the before and after:
```
String Stream:
Benchmarking rebuild... 1464148 uuids/sec
Iteration:
Benchmarking rebuild... 17086712 uuids/sec
```