@vihangm 2021-03-07
Realized that the rebuild benchmark I added was timing both the call to `str` and the `rebuild`. Fixed it to stringify outside the benchmark and the improvement now seems to line up with what google benchmark reports.